📚  Refreshed the Docs
These docs were a nice format, from another project, so we started
there, and this update pulls them into line with this Client code,
insofar as we have been able to test. Several things are still not
tested, and we are working with Peach to get those hammered out, but at
least everything is in place, and should be a lot better than before.