📦  Added Payment Instruments and Testing
This concludes the bare minimum of the utility we need from the Client,
and all key functions are tested in the scripts. This is a big chunk of
work because we found we needed to create Payment Instruments even if we
are doing all the payment processing... so that had to be added and
tested.