🧯  Found another error location for Peach
It turns out there is another filure reporting scheme from Peach - one
that's not documented as far as I could read, and now all the calls look
for that error as well. If it's there, then we'll create an error
message from it, and that will really help the debugging when something
goes wrong.